requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
facenet-pytorch